change showId type from int to Long

+++ app/src/main/java/br/com/locaweb/jacatv/fragment/CastListFragment.java
@@ -23,7 +23,7 @@ public class CastListFragment extends ListFragment {
     CastAdapter adapter;
 
     @FragmentArg
-    int showId;
+    Long showId;
 
     @RestService
     RestConnection connection;
